Resolves failing Cucumber tests and adds a development environment  (#138)
* Added development environment

This adds a `dev/start` and `dev/stop` scripts to handle starting and
stopping a local development environment.

* Corrected scope for validating an added host

Once v5 scope was restricted to ensure a resource couldn't see all
available resources, it broke these tests. This change uses the admin
user connection, which has permission to view the created resource,
to validate the resource's presence.

* Fixes to Docker Compose file & Docs

This commit includes updates to the docker-compose.yml & README file
which allow tests to run using the mounted authentication socket.