Utils: Do not force pull in test.sh (#146)
test.sh used to do `docker-compose pull` which unconditionally pulls
fresh images. It's often a waste of bandwidth and time:
`docker-compose build` already pulls any missing images, whereas
updating on every run means cache is busted and many images potentially
need to be rebuilt.

OTOH in CI we really want to make sure the tests pass against the
latest images; otherwise we can have misleading results which are
hard to debug.

So try to pull, but ignore failures. This makes sure the pipelines
are robust in face of network failure and enable offline usage with
warm cache, but still ensures the test targets are fresh if possible.