Add ability to add a chained cert in cert_utils
`OpenSSL::X509::Store.new.add_cert` adds only the intermediate certificate
to the store. When adding a certificate to a store we (or anyone consuming
this api) can use cert_utils's `add_chained_cert` method.
If the certificate has more than one certificate in its chain, it will be parsed and added to the store one by one.