Custom rest_client_options via Conjur configuration
This replaces the monkey patch to RestClient::Request that made it always uses :ssl_cert_store, which polluted into usage of RestClient outside the scope of Conjur. The new approach provides a sane default that retains the previous behavior, and gives the user the ability to configure RestClient::Requests scoped only to Conjur.